+++ sdrgui/gui/editcommanddialog.cpp
@@ -204,7 +204,7 @@ void EditCommandDialog::setKeyLabel()
     else if (m_keyModifiers != Qt::NoModifier)
     {
         QString altGrStr = m_keyModifiers & Qt::GroupSwitchModifier ? "Gr " : "";
-        int maskedModifiers = (m_keyModifiers & 0x3FFFFFFF) + ((m_keyModifiers & 0x40000000)>>3);
+        int maskedModifiers = ((int) m_keyModifiers & 0x3FFFFFFF) + (((int) m_keyModifiers & 0x40000000)>>3);
         ui->keyLabel->setText(altGrStr + QKeySequence(maskedModifiers, m_key).toString());
     }
     else
